\name{hsaXXXXX}
\alias{hsaXXXXX}
\alias{hsa00010}
\alias{hsa00020}
\alias{hsa00030}
\alias{hsa00040}
\alias{hsa00051}
\alias{hsa00052}
\alias{hsa00053}
\alias{hsa00061}
\alias{hsa00062}
\alias{hsa00071}
\alias{hsa00072}
\alias{hsa00100}
\alias{hsa00120}
\alias{hsa00130}
\alias{hsa00140}
\alias{hsa00190}
\alias{hsa00230}
\alias{hsa00232}
\alias{hsa00240}
\alias{hsa00250}
\alias{hsa00260}
\alias{hsa00270}
\alias{hsa00280}
\alias{hsa00290}
\alias{hsa00300}
\alias{hsa00310}
\alias{hsa00330}
\alias{hsa00340}
\alias{hsa00350}
\alias{hsa00360}
\alias{hsa00380}
\alias{hsa00400}
\alias{hsa00410}
\alias{hsa00430}
\alias{hsa00450}
\alias{hsa00460}
\alias{hsa00471}
\alias{hsa00472}
\alias{hsa00480}
\alias{hsa00500}
\alias{hsa00510}
\alias{hsa00511}
\alias{hsa00512}
\alias{hsa00514}
\alias{hsa00520}
\alias{hsa00524}
\alias{hsa00531}
\alias{hsa00532}
\alias{hsa00533}
\alias{hsa00534}
\alias{hsa00561}
\alias{hsa00562}
\alias{hsa00563}
\alias{hsa00564}
\alias{hsa00565}
\alias{hsa00590}
\alias{hsa00591}
\alias{hsa00592}
\alias{hsa00600}
\alias{hsa00601}
\alias{hsa00603}
\alias{hsa00604}
\alias{hsa00620}
\alias{hsa00630}
\alias{hsa00640}
\alias{hsa00650}
\alias{hsa00670}
\alias{hsa00730}
\alias{hsa00740}
\alias{hsa00750}
\alias{hsa00760}
\alias{hsa00770}
\alias{hsa00780}
\alias{hsa00785}
\alias{hsa00790}
\alias{hsa00830}
\alias{hsa00860}
\alias{hsa00900}
\alias{hsa00910}
\alias{hsa00920}
\alias{hsa00970}
\alias{hsa00980}
\alias{hsa00982}
\alias{hsa00983}
\alias{hsa01040}
\alias{hsa01100}
\alias{hsa02010}
\alias{hsa03010}
\alias{hsa03013}
\alias{hsa03015}
\alias{hsa03018}
\alias{hsa03020}
\alias{hsa03022}
\alias{hsa03030}
\alias{hsa03040}
\alias{hsa03050}
\alias{hsa03060}
\alias{hsa03320}
\alias{hsa03410}
\alias{hsa03420}
\alias{hsa03430}
\alias{hsa03440}
\alias{hsa03450}
\alias{hsa04010}
\alias{hsa04012}
\alias{hsa04020}
\alias{hsa04060}
\alias{hsa04062}
\alias{hsa04070}
\alias{hsa04080}
\alias{hsa04110}
\alias{hsa04114}
\alias{hsa04115}
\alias{hsa04120}
\alias{hsa04122}
\alias{hsa04130}
\alias{hsa04140}
\alias{hsa04141}
\alias{hsa04142}
\alias{hsa04144}
\alias{hsa04145}
\alias{hsa04146}
\alias{hsa04150}
\alias{hsa04210}
\alias{hsa04260}
\alias{hsa04270}
\alias{hsa04310}
\alias{hsa04320}
\alias{hsa04330}
\alias{hsa04340}
\alias{hsa04350}
\alias{hsa04360}
\alias{hsa04370}
\alias{hsa04380}
\alias{hsa04510}
\alias{hsa04512}
\alias{hsa04514}
\alias{hsa04520}
\alias{hsa04530}
\alias{hsa04540}
\alias{hsa04610}
\alias{hsa04612}
\alias{hsa04614}
\alias{hsa04620}
\alias{hsa04621}
\alias{hsa04622}
\alias{hsa04623}
\alias{hsa04630}
\alias{hsa04640}
\alias{hsa04650}
\alias{hsa04660}
\alias{hsa04662}
\alias{hsa04664}
\alias{hsa04666}
\alias{hsa04670}
\alias{hsa04672}
\alias{hsa04710}
\alias{hsa04720}
\alias{hsa04722}
\alias{hsa04730}
\alias{hsa04740}
\alias{hsa04742}
\alias{hsa04744}
\alias{hsa04810}
\alias{hsa04910}
\alias{hsa04912}
\alias{hsa04914}
\alias{hsa04916}
\alias{hsa04920}
\alias{hsa04930}
\alias{hsa04940}
\alias{hsa04950}
\alias{hsa04960}
\alias{hsa04962}
\alias{hsa04964}
\alias{hsa04966}
\alias{hsa04970}
\alias{hsa04971}
\alias{hsa04972}
\alias{hsa04973}
\alias{hsa04974}
\alias{hsa04975}
\alias{hsa04976}
\alias{hsa05010}
\alias{hsa05012}
\alias{hsa05014}
\alias{hsa05016}
\alias{hsa05020}
\alias{hsa05100}
\alias{hsa05110}
\alias{hsa05120}
\alias{hsa05130}
\alias{hsa05131}
\alias{hsa05140}
\alias{hsa05142}
\alias{hsa05144}
\alias{hsa05145}
\alias{hsa05146}
\alias{hsa05150}
\alias{hsa05160}
\alias{hsa05200}
\alias{hsa05210}
\alias{hsa05211}
\alias{hsa05212}
\alias{hsa05213}
\alias{hsa05214}
\alias{hsa05215}
\alias{hsa05216}
\alias{hsa05217}
\alias{hsa05218}
\alias{hsa05219}
\alias{hsa05220}
\alias{hsa05221}
\alias{hsa05222}
\alias{hsa05223}
\alias{hsa05310}
\alias{hsa05320}
\alias{hsa05322}
\alias{hsa05330}
\alias{hsa05332}
\alias{hsa05340}
\alias{hsa05410}
\alias{hsa05412}
\alias{hsa05414}
\alias{hsa05416}
\docType{data}
\title{
KEGG pathways
}
\description{
Each file "hsaXXXXX" contains the gene indetifiers of the pathway
}
\usage{data(hsa00010)}
\format{
 10327 124 125 126 127 ...
}
\section{Pathways:}{
\preformatted{
hsa00010,hsa00020,hsa00030,hsa00040,hsa00051,hsa00052,hsa00053,hsa00061,
hsa00062,hsa00071,hsa00072,hsa00100,hsa00120,hsa00130,hsa00140,hsa00190,
hsa00230,hsa00232,hsa00240,hsa00250,hsa00260,hsa00270,hsa00280,hsa00290,
hsa00300,hsa00310,hsa00330,hsa00340,hsa00350,hsa00360,hsa00380,hsa00400,
hsa00410,hsa00430,hsa00450,hsa00460,hsa00471,hsa00472,hsa00480,hsa00500,
hsa00510,hsa00511,hsa00512,hsa00514,hsa00520,hsa00524,hsa00531,hsa00532,
hsa00533,hsa00534,hsa00561,hsa00562,hsa00563,hsa00564,hsa00565,hsa00590,
hsa00591,hsa00592,hsa00600,hsa00601,hsa00603,hsa00604,hsa00620,hsa00630,
hsa00640,hsa00650,hsa00670,hsa00730,hsa00740,hsa00750,hsa00760,hsa00770,
hsa00780,hsa00785,hsa00790,hsa00830,hsa00860,hsa00900,hsa00910,hsa00920,
hsa00970,hsa00980,hsa00982,hsa00983,hsa01040,hsa01100,hsa02010,hsa03010,
hsa03013,hsa03015,hsa03018,hsa03020,hsa03022,hsa03030,hsa03040,hsa03050,
hsa03060,hsa03320,hsa03410,hsa03420,hsa03430,hsa03440,hsa03450,hsa04010,
hsa04012,hsa04020,hsa04060,hsa04062,hsa04070,hsa04080,hsa04110,hsa04114,
hsa04115,hsa04120,hsa04122,hsa04130,hsa04140,hsa04141,hsa04142,hsa04144,
hsa04145,hsa04146,hsa04150,hsa04210,hsa04260,hsa04270,hsa04310,hsa04320,
hsa04330,hsa04340,hsa04350,hsa04360,hsa04370,hsa04380,hsa04510,hsa04512,
hsa04514,hsa04520,hsa04530,hsa04540,hsa04610,hsa04612,hsa04614,hsa04620,
hsa04621,hsa04622,hsa04623,hsa04630,hsa04640,hsa04650,hsa04660,hsa04662,
hsa04664,hsa04666,hsa04670,hsa04672,hsa04710,hsa04720,hsa04722,hsa04730,
hsa04740,hsa04742,hsa04744,hsa04810,hsa04910,hsa04912,hsa04914,hsa04916,
hsa04920,hsa04930,hsa04940,hsa04950,hsa04960,hsa04962,hsa04964,hsa04966,
hsa04970,hsa04971,hsa04972,hsa04973,hsa04974,hsa04975,hsa04976,hsa05010,
hsa05012,hsa05014,hsa05016,hsa05020,hsa05100,hsa05110,hsa05120,hsa05130,
hsa05131,hsa05140,hsa05142,hsa05144,hsa05145,hsa05146,hsa05150,hsa05160,
hsa05200,hsa05210,hsa05211,hsa05212,hsa05213,hsa05214,hsa05215,hsa05216,
hsa05217,hsa05218,hsa05219,hsa05220,hsa05221,hsa05222,hsa05223,hsa05310,
hsa05320,hsa05322,hsa05330,hsa05332,hsa05340,hsa05410,hsa05412,hsa05414,
hsa05416
}
}

\source{
http://www.genome.jp/kegg/
}
\examples{
data(hsa00010)
data(hsa00010,hsa00020,hsa00030,hsa00040,hsa04144,hsa04510,hsa04740,hsa04810)
}
\keyword{datasets}
