\name{RHSAXXXXX}
\alias{RHSAXXXXX}
\alias{RHSA109582}
\alias{RHSA164843}
\alias{RHSA446343}
\alias{RHSA1474244}
\alias{RHSA8876384}
\alias{RHSA8964572}
\docType{data}
\title{
Reactome Pathway examples
}
\description{
Each file "RHSAXXXX" contains the gene identifiers.
}
\usage{data(RHSA164843)}
\format{
  The format is:
 num [1:6] 11168 155030 155348 155459 155908   2547...
}
\source{
reactome.db
}
\examples{
data(RHSA164843)
}
\keyword{datasets}
