\name{hsaXXXXX}
\alias{hsaXXXXX}
\alias{hsa00100}
\alias{hsa00120}
\alias{hsa00130}
\alias{hsa00140}
\alias{hsa00190}
\alias{hsa02010}
\docType{data}
\title{
KEGG pathways examples
}
\description{
Each file "hsaXXXXX" contains the gene indetifiers of the pathway
}
\usage{data(hsa02010)}
\format{
 10327 124 125 126 127 ...
}
\section{Pathways:}{
\preformatted{
hsa00100,hsa00120,hsa00130,hsa00140,hsa00190,hsa02010
}
}

\source{
http://www.genome.jp/kegg/
}
\examples{
\dontrun{
data(hsa02010)
data(hsa00100,hsa00120,hsa00130,hsa00140,hsa00190,hsa02010)
}
}
\keyword{datasets}
